% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csortb-package.r
\docType{package}
\name{csortb}
\alias{csortb}
\alias{csortb-package}
\title{csortb.}
\description{
Residential Tenancies Board (RTB) Average Monthly Rent Report by
 Property Type, Location, Number of Bedrooms and Year
 2008 to 2016 hosted by Ireland's Central Statistics Offices
}
